cleaned and extracted cartItemsWidget
